+++ packages/react/src/hooks/api-session.hook.ts
@@ -2,6 +2,7 @@ import { useMemo } from 'react';
 import { useAuthContext } from '../contexts/auth.context';
 import { Session } from '../definitions/session';
 import { useAuth } from './auth.hook';
+import { AuthWalletType } from '../definitions/auth';
 
 export interface ApiSessionInterface {
   isInitialized: boolean;
@@ -16,6 +17,8 @@ export interface ApiSessionInterface {
     discount?: string,
     wallet?: string,
     ref?: string,
+    walletType?: AuthWalletType,
+    recommendationCode?: string,
   ) => Promise<string>;
   createSessionNew: (
     address: string,
@@ -24,6 +27,8 @@ export interface ApiSessionInterface {
     discount?: string,
     wallet?: string,
     ref?: string,
+    walletType?: AuthWalletType,
+    recommendationCode?: string,
   ) => Promise<string>;
   updateSession: (token: string) => void;
   deleteSession: () => Promise<void>;
@@ -41,9 +46,13 @@ export function useApiSession(): ApiSessionInterface {
     discount?: string,
     wallet?: string,
     ref?: string,
+    walletType?: AuthWalletType,
+    recommendationCode?: string,
   ): Promise<string> {
     return (
-      isSignUp ? signUp(address, signature, key, discount, wallet, ref) : signIn(address, signature, key, discount)
+      isSignUp
+        ? signUp(address, signature, key, discount, wallet, ref, walletType, recommendationCode)
+        : signIn(address, signature, key, discount, walletType)
     ).then(({ accessToken }) => {
       setAuthToken(accessToken);
       return accessToken;
@@ -57,11 +66,15 @@ export function useApiSession(): ApiSessionInterface {
     discount?: string,
     wallet?: string,
     ref?: string,
+    walletType?: AuthWalletType,
+    recommendationCode?: string,
   ) {
-    return authenticate(address, signature, key, discount, wallet, ref).then(({ accessToken }) => {
-      setAuthToken(accessToken);
-      return accessToken;
-    });
+    return authenticate(address, signature, key, discount, wallet, ref, walletType, recommendationCode).then(
+      ({ accessToken }) => {
+        setAuthToken(accessToken);
+        return accessToken;
+      },
+    );
   }
 
   function updateSession(token: string) {