+++ packages/react/src/hooks/kyc.hook.ts
@@ -24,6 +24,8 @@ import {
   KycBeneficialData,
   KycOperationalData,
   PaymentData,
+  RecallData,
+  KycRecommendationData,
 } from '../definitions/kyc';
 import { useApi } from './api.hook';
 
@@ -50,7 +52,9 @@ export interface KycInterface {
   setPersonalData: (code: string, url: string, data: KycPersonalData) => Promise<KycStepBase>;
   setManualIdentData: (code: string, url: string, data: KycManualIdentData) => Promise<KycStepBase>;
   setLegalEntityData: (code: string, url: string, data: KycLegalEntityData) => Promise<KycStepBase>;
+  setSoleProprietorshipData: (code: string, url: string, data: KycFileData) => Promise<KycStepBase>;
   setNationalityData: (code: string, url: string, data: KycNationalityData) => Promise<KycStepBase>;
+  setRecommendationData: (code: string, url: string, data: KycRecommendationData) => Promise<KycStepBase>;
   setFileData: (code: string, url: string, data: KycFileData) => Promise<KycStepBase>;
   setSignatoryPowerData: (code: string, url: string, data: KycSignatoryPowerData) => Promise<KycStepBase>;
   setBeneficialData: (code: string, url: string, data: KycBeneficialData) => Promise<KycStepBase>;
@@ -58,6 +62,7 @@ export interface KycInterface {
   getFinancialData: (code: string, url: string, lang?: string) => Promise<KycFinancialQuestions>;
   setFinancialData: (code: string, url: string, data: KycFinancialResponses) => Promise<KycStepBase>;
   setPaymentData: (code: string, url: string, data: PaymentData) => Promise<KycStepBase>;
+  setRecallData: (code: string, url: string, data: RecallData) => Promise<KycStepBase>;
   getFile: (kycFileId: string) => Promise<KycFile>;
 
   // 2fa
@@ -132,7 +137,15 @@ export function useKyc(): KycInterface {
     return call({ url, code, method: 'PUT', data });
   }
 
+  function setSoleProprietorshipData(code: string, url: string, data: KycFileData): Promise<KycStepBase> {
+    return call({ url, code, method: 'PUT', data });
+  }
+
   async function setNationalityData(code: string, url: string, data: KycNationalityData): Promise<KycStepBase> {
+    return call({ url, code, method: 'PUT', data });
+  }
+
+  async function setRecommendationData(code: string, url: string, data: KycRecommendationData): Promise<KycStepBase> {
     return call({ url, code, method: 'PUT', data });
   }
 
@@ -158,6 +171,10 @@ export function useKyc(): KycInterface {
   }
 
   async function setPaymentData(code: string, url: string, data: PaymentData): Promise<KycStepBase> {
+    return call({ url, code, method: 'PUT', data });
+  }
+
+  async function setRecallData(code: string, url: string, data: RecallData): Promise<KycStepBase> {
     return call({ url, code, method: 'PUT', data });
   }
 
@@ -233,7 +250,9 @@ export function useKyc(): KycInterface {
       setPersonalData,
       setManualIdentData,
       setLegalEntityData,
+      setSoleProprietorshipData,
       setNationalityData,
+      setRecommendationData,
       setFileData,
       setSignatoryPowerData,
       setBeneficialData,
@@ -242,6 +261,7 @@ export function useKyc(): KycInterface {
       getFile,
       setFinancialData,
       setPaymentData,
+      setRecallData,
       check2fa,
       setup2fa,
       verify2fa,